requirements.txt
asgiref==3.8.1
Django==5.0.6
lxml==5.2.2
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.4
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
gunicorn==22.0.0